+++ Source/SpaceHorror/Public/FireMechanicAuto.h
@@ -7,6 +7,8 @@
 #include "FireMechanicAuto.generated.h"
 
 
+class AMasterWeapons;
+
 UCLASS( ClassGroup=(Custom), meta=(BlueprintSpawnableComponent) )
 class SPACEHORROR_API UFireMechanicAuto : public UActorComponent
 {
@@ -21,9 +23,24 @@ protected:
 	virtual void BeginPlay() override;
 
 public:	
+
+	AMasterWeapons* MasterWeapons = nullptr;
 	// Called every frame
 	virtual void TickComponent(float DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction) override;
 
 	bool IsHoldingThisWeapon;
+
+	void Fire();
+
+	bool canFire;
+
+	bool IsHoldingTrigger;
+
+	bool IsPressFire;
+
+	float fireRate;
+
+	float firedTimeCount = 0;
+	
 		
 };